+++ packages/psychart/src/pumpchart/pumpchart.ts
@@ -1,9 +1,10 @@
 import * as SMath from 'smath';
 import { Chart } from '../chart';
 import { defaultPumpchartDataOptions, defaultPumpchartOptions } from './defaults';
-import { Point, PumpchartDataOptions, PumpchartOptions, State } from './types';
+import { Point, PumpchartDataOptions, PumpchartOptions, PumpchartState } from './types';
 import { Color, Palette } from 'viridis';
 import { TextAnchor } from '../types';
+import { f, toFunction, zero } from './lib';
 
 /**
  * Show a pump's relationship between flow rate and pressure at different operating conditions.
@@ -13,12 +14,20 @@ export class Pumpchart extends Chart<PumpchartOptions> {
      * Layers of the SVG as groups.
      */
     private readonly g = {
-        axes: document.createElementNS(this.NS, 'g'),
         curves: document.createElementNS(this.NS, 'g'),
+        axes: document.createElementNS(this.NS, 'g'),
         data: document.createElementNS(this.NS, 'g'),
         text: document.createElementNS(this.NS, 'g'),
         tips: document.createElementNS(this.NS, 'g'),
     };
+    /**
+     * The pump curve `h = p(q)`
+     */
+    private readonly p: f;
+    /**
+     * The system curve `h = s(q)`
+     */
+    private readonly s: f;
     /**
      * The maximum flow rate shown on the x-axis
      */
@@ -48,9 +57,13 @@ export class Pumpchart extends Chart<PumpchartOptions> {
         super(options, defaultPumpchartOptions);
         // Append all groups to the SVG.
         Object.values(this.g).forEach(group => this.svg.appendChild(group));
+        // Initialize the pump curve and system curve
+        this.p = toFunction(this.options.curve.pump, 'q');
+        this.s = toFunction(this.options.curve.system, 'q');
         // Compute the axes limits and intervals
-        this.maxFlow = this.options.pumpMaxFlow * 1.1;
-        this.maxHead = this.options.pumpMaxHead * 1.1;
+        const q0: number = zero(this.p, 0, 1e6); // Max pump flow with 0 head
+        this.maxFlow = 1.1 * q0;
+        this.maxHead = 1.1 * this.p(0);
         const flowStep: number = Pumpchart.getStep(this.maxFlow, this.options.size.x / this.options.font.size / 6)
         const headStep: number = Pumpchart.getStep(this.maxHead, this.options.size.y / this.options.font.size / 3)
         // Create the axes.
@@ -95,15 +108,46 @@ export class Pumpchart extends Chart<PumpchartOptions> {
             // Show axis label
             this.drawLabel(`${head}${this.options.units.head}`, { flow: 0, head: head }, TextAnchor.E, `Head [${this.options.units.head}]`);
         }
-        this.drawPerformanceCurves();
-        this.drawSystemCurveAndOperation();
+        // Draw the system curve
+        const sysColor: Color = Color.hex(this.options.systemCuveColor);
+        const n: number = SMath.clamp(SMath.normalize(this.options.speed.operation, 0, this.options.speed.max), 0.01, 1);
+        const p: f = this.scale(this.p, n);
+        const qmax: number = zero(q => this.s(q) - this.p(q), 0, 1e6); // flow @ max speed
+        const qop: number = zero(q => this.s(q) - p(q), 0, 1e6); // flow @ operation
+        const opPt: PumpchartState = { flow: qop, head: p(qop), speed: this.options.speed.operation };
+        this.drawCurve('System Curve', sysColor, 2 * this.options.axisWidth, this.s, 0, qmax);
+        // Draw operation axis lines
+        const operation = this.createPath([
+            { flow: 0, head: opPt.head },
+            opPt,
+            { flow: opPt.flow, head: 0 },
+        ], false);
+        operation.setAttribute('fill', 'none');
+        operation.setAttribute('stroke', sysColor.toString());
+        operation.setAttribute('stroke-width', `${this.options.axisWidth}px`);
+        operation.setAttribute('stroke-linecap', 'round');
+        this.g.curves.append(operation);
+        // Draw the operating point
+        this.plot(opPt, {
+            name: 'Operation Point',
+            radius: 5 * this.options.axisWidth,
+            color: this.options.systemCuveColor,
+        });
+        // Draw concentric pump performance curves
+        const pumpColor: Color = Color.hex(this.options.pumpCurveColor);
+        this.drawCurve(`Performance Curve at ${SMath.round2(this.options.speed.max, 0.1)}${this.options.units.speed}`, pumpColor, this.options.axisWidth * 2, this.p, 0, q0);
+        this.options.speed.steps.forEach(speed => {
+            const pct: number = SMath.clamp(SMath.normalize(speed, 0, this.options.speed.max), 0.01, 1);
+            const p1: f = this.scale(this.p, pct);
+            this.drawCurve('', pumpColor, this.options.axisWidth, p1, 0, q0 * pct);
+        });
     }
     /**
      * Convert a state to an (x,y) coordinate.
      * @param state Any state in this system
      * @returns An (x,y) coordinate on the screen
      */
-    private state2xy(state: State): Point {
+    private state2xy(state: PumpchartState): Point {
         const xMin: number = this.options.padding.x;
         const xMax: number = this.options.size.x - this.options.padding.x
         const yMin: number = this.options.padding.y;
@@ -119,7 +163,7 @@ export class Pumpchart extends Chart<PumpchartOptions> {
      * @param closePath Whether or not to close the path
      * @returns A `<path>` element containing the array of states
      */
-    private createPath(data: State[], closePath = false): SVGPathElement {
+    private createPath(data: PumpchartState[], closePath = false): SVGPathElement {
         const path: SVGPathElement = document.createElementNS(this.NS, 'path');
         path.setAttribute('d', 'M ' + data.map(pt => {
             const xy: Point = this.state2xy(pt);
@@ -135,7 +179,7 @@ export class Pumpchart extends Chart<PumpchartOptions> {
      * @param anchor Label text anchor
      * @param tooltip Optional tooltip text on mouse hover
      */
-    private drawLabel(content: string, location: State, anchor: TextAnchor, tooltip?: string): void {
+    private drawLabel(content: string, location: PumpchartState, anchor: TextAnchor, tooltip?: string): void {
         const axisColor: Color = Color.hex(this.options.axisColor);
         const label: SVGTextElement = this.createLabel(content, this.state2xy(location), axisColor, anchor);
         this.g.text.appendChild(label);
@@ -148,7 +192,7 @@ export class Pumpchart extends Chart<PumpchartOptions> {
      * Draw a curve `h = f(q)` on the curves layer.
      */
     private drawCurve(tooltip: string, color: Color, width: number, h: f, min = 0, max = this.maxFlow, steps = 1e3): void {
-        const states: State[] = SMath.linspace(min, max, steps).map<State>(q => { return { flow: q, head: h(q) } });
+        const states: PumpchartState[] = SMath.linspace(min, max, steps).map<PumpchartState>(q => { return { flow: q, head: h(q) } });
         const curve: SVGPathElement = this.createPath(states, false);
         curve.setAttribute('fill', 'none');
         curve.setAttribute('stroke', color.toString());
@@ -163,7 +207,7 @@ export class Pumpchart extends Chart<PumpchartOptions> {
     /**
      * Draw a custom circle onto any layer.
      */
-    private drawCircle(tooltip: string, color: Color, state: State, radius: number, parent: SVGElement): void {
+    private drawCircle(tooltip: string, color: Color, state: PumpchartState, radius: number, parent: SVGElement): void {
         const circle: SVGCircleElement = document.createElementNS(this.NS, 'circle');
         const center: Point = this.state2xy(state);
         circle.setAttribute('fill', color.toString());
@@ -177,82 +221,34 @@ export class Pumpchart extends Chart<PumpchartOptions> {
         }
     }
     /**
-     * Draw concentric pump performance curves.
+     * Scale a function.
      */
-    private drawPerformanceCurves(): void {
-        const color: Color = Color.hex(this.options.performanceCurveColor);
-        const p: f = this.performanceCurve();
-        const q0: number = this.options.pumpMaxFlow;
-        const minor: number = this.options.performanceSteps;
-        this.drawCurve('Performance Curve', color, 2, p, 0, q0);
-        if (minor > 0) {
-            const minorCurvePoints: number[] = SMath.linspace(0, 1, minor + 1);
-            minorCurvePoints.shift(); // Cut off the 0%
-            minorCurvePoints.pop(); // Cut off the 100%
-            minorCurvePoints.forEach(pct => {
-                const p1: f = this.performanceCurve(pct);
-                this.drawCurve('', color, 1, p1, 0, q0 * pct);
-            });
-        }
-    }
-    /**
-     * Draw the system curve and operational point.
-     */
-    private drawSystemCurveAndOperation(): void {
-        const color: Color = Color.hex(this.options.systemCuveColor);
-        const q0: number = this.options.systemOpFlow;
-        const p: f = this.performanceCurve();
-        const s: f = this.systemCurve();
-        this.drawCurve('System Curve', color, 2, s, 0, q0);
-        // Draw operation axis lines
-        const operation = this.createPath([
-            { flow: 0, head: p(q0) },
-            { flow: q0, head: p(q0) },
-            { flow: q0, head: 0 },
-        ], false);
-        operation.setAttribute('fill', 'none');
-        operation.setAttribute('stroke', color.toString());
-        operation.setAttribute('stroke-width', '1px');
-        operation.setAttribute('stroke-linecap', 'round');
-        this.g.curves.append(operation);
-        // Draw operating point
-        this.drawCircle(`Operation Point\nFlow = ${q0.toFixed(1)} [${this.options.units.flow}]\nHead = ${p(q0).toFixed(1)} [${this.options.units.head}]`, color, { flow: q0, head: p(q0) }, 5, this.g.curves);
-    }
-    /**
-     * Generate the pump performance curve.
-     * @param rpmPct The normalized speed percent
-     * @returns The performance curve `h = p(q)`
-     */
-    private performanceCurve(rpmPct = 1): f {
-        const h0: number = this.options.pumpMaxHead * rpmPct;
-        const q0: number = this.options.pumpMaxFlow * rpmPct;
-        return q => h0 * (1 - (q / q0) ** 2);
-    }
-    /**
-     * Generate the system curve.
-     * @returns The system curve `h = s(q)`
-     */
-    private systemCurve(): f {
-        const p: f = this.performanceCurve();
-        const h0: number = this.options.systemMinHead;
-        const q0: number = this.options.systemOpFlow;
-        return q => h0 + (p(q0) - h0) * (q / q0) ** 2;
+    private scale(p: f, n: number): f {
+        return q => n * p(q / n);
     }
     /**
      * Plot a single data point.
      * @param state The current state of the system
      * @param config Display options for plotting data
      */
-    public plot(state: State, config: Partial<PumpchartDataOptions> = {}): void {
+    public plot(state: PumpchartState, config: Partial<PumpchartDataOptions> = {}): void {
         const options: PumpchartDataOptions = Chart.setDefaults(config, defaultPumpchartDataOptions);
-        const color: Color = Palette[options.gradient.name].getColor(options.gradient.score);
+        const color: Color = options.timestamp > 0 ? Palette[this.options.gradient].getColor(options.timestamp, this.options.timestamp.start, this.options.timestamp.stop) : Color.hex(options.color);
+        const speedEstimator: f = n => this.scale(this.p, n)(state.flow) - state.head;
+        let speedEstimate: number;
+        try {
+            speedEstimate = SMath.expand(zero(speedEstimator, 0.01, 1), 0, this.options.speed.max);
+        } catch {
+            speedEstimate = 0;
+        }
         const tip: string =
             (options.name ? `${options.name}\n` : '') +
             (options.timestamp > 0 ? `${new Date(options.timestamp).toLocaleString()}\n` : '') +
-            `Flow = ${state.flow} [${this.options.units.flow}]` +
-            `\nHead = ${state.head} [${this.options.units.head}]` +
-            (typeof state.speed !== 'undefined' ? `\nSpeed = ${state.speed} [${this.options.units.speed}]` : '') +
-            (typeof state.power !== 'undefined' ? `\nPower = ${state.power} [${this.options.units.power}]` : '');
+            `Flow = ${SMath.round2(state.flow, 0.1)}${this.options.units.flow}` +
+            `\nHead = ${SMath.round2(state.head, 0.1)}${this.options.units.head}` +
+            (speedEstimate > 0 ? `\nSpeed = ${SMath.round2(speedEstimate, 0.1)}${this.options.units.speed} (est.)` : '\nSpeed cannot be estimated') +
+            (typeof state.speed !== 'undefined' ? `\nSpeed = ${SMath.round2(state.speed, 0.1)}${this.options.units.speed}` : '') +
+            (typeof state.power !== 'undefined' ? `\nPower = ${SMath.round2(state.power, 0.1)}${this.options.units.power}` : '');
         this.drawCircle(tip, color, state, options.radius, this.g.data);
     }
     /**
@@ -262,8 +258,3 @@ export class Pumpchart extends Chart<PumpchartOptions> {
         Chart.clearChildren(this.g.data);
     }
 }
-
-/**
- * Represents a mathematical function `h = f(q)`
- */
-type f = (q: number) => number;